+++ app/page.tsx
@@ -1,8 +1,8 @@
-import { Home } from "./components/sections/Home";
-import { About } from "./components/sections/About";
-import { Resume } from "./components/sections/Resume";
-import { Portfolio } from "./components/sections/Portfolio";
-import { Contact } from "./components/sections/Contact";
+import { Home } from "./sections/Home";
+import { About } from "./sections/About";
+import { Resume } from "./sections/Resume";
+import { Portfolio } from "./sections/Portfolio";
+import { Contact } from "./sections/Contact";
 
 export default function Page() {
   return (
